front: extract clear button logic into new clearButton component
